Replace common git impl with `sdk-sync`'s impl
