+++ tools/smithy-rs-tool-common/fake-cli/git-changed-files-empty
