+++ tools/smithy-rs-tool-common/fake-cli/git-current-branch-name
