+++ tools/smithy-rs-tool-common/fake-cli/git-delete-branch
