+++ tools/smithy-rs-tool-common/fake-cli/git-extract-commit-info
