+++ tools/smithy-rs-tool-common/fake-cli/git-get-head-revision
