+++ tools/smithy-rs-tool-common/fake-cli/git-rev-list-path
