+++ tools/smithy-rs-tool-common/fake-cli/git-show
