+++ tools/smithy-rs-tool-common/fake-cli/git-squash-merge
