+++ tools/smithy-rs-tool-common/fake-cli/git-untracked-files-empty
