--- tools/smithy-rs-tool-common/fake_git/git_fails
@@ -1,4 +1,0 @@
-#!/bin/bash
-echo "some stdout failure message"
->&2 echo "some stderr failure message"
-exit 1
