--- tools/smithy-rs-tool-common/fake_git/git_revparse_show_toplevel
@@ -1,6 +1,0 @@
-#!/bin/bash
-if [[ "$1" != "rev-parse" || "$2" != "--show-toplevel" ]]; then
-    echo "wrong arguments" >&2
-    exit 1
-fi
-echo "/git/repo/root/path"
